Fixes to the basic information controls
Combined the essential info and the description into one view. Also, added a valid survey check before saving to the database.
