+++ SurveyManager/forms/surveyMenu/basicInfo/InfoCtl.resx
