+++ SurveyManager/forms/surveyMenu/basicInfo/LocationCtl.resx
